Compute 0.3

+++ popcorn.php
@@ -39,7 +39,7 @@
               print("<tr>");
               $productName = (fgets($fh));
               print("<td>".$productName."</td>");
-              $productPrice = (int)(fgets($fh));
+              $productPrice = (float)(fgets($fh));
               printf("<td>$ %4.2f</td>", $productPrice);
               print("<td align = \"center\"><input type = \"text\" name = \"product".$i."\" size = \"3\" /></td>");
               print("</tr>");
